新增发布新文章功能demo

+++ src/main/java/com/murasame/entity/Blogs.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class Blogs {
-	BigInteger id;
+	Long id;
 	Integer author_id;
 	LocalDateTime created_at;
 	LocalDateTime updated_at;
