+++ src/main/java/com/murasame/entity/Comments.java
@@ -7,9 +7,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class Comments {
-	BigInteger id;
-	BigInteger blog_id;
-	BigInteger parent_cid;
+	Long id;
+	Long blog_id;
+	Long parent_cid;
 	Integer author_id;
 	String content;
 	LocalDateTime created_at;
